Add recaptcha.enabled property so that if recaptcha is not configured, the new user form will still work. By default recaptcha is not enabled.
