+++ account-management-app/src/test/java/org/duracloud/account/app/controller/UserControllerTest.java
@@ -198,7 +198,7 @@ public class UserControllerTest extends AmaControllerTestBase {
         Assert.assertEquals(UserController.USER_EDIT_VIEW, mav.getViewName());
     }
 
-    //@Test
+    @Test
     public void testAddUser() throws Exception {
         EasyMock.expect(userService.createNewUser(EasyMock.isA(String.class),
                                                   EasyMock.isA(String.class),
@@ -231,15 +231,15 @@ public class UserControllerTest extends AmaControllerTestBase {
 
         replayMocks();
 
-        ModelAndView mav = userController.add(newUserForm, result, null, null, null);
+        ModelAndView mav = userController.add(newUserForm, result, null, null);
         Assert.assertNotNull(mav);
     }
 
-    //@Test
+    @Test
     public void testAddUserErrors() throws Exception {
         setupHasBindingResultErrors(true);
         replayMocks();
-        ModelAndView mav = userController.add(null, result, model, null, null);
+        ModelAndView mav = userController.add(null, result, model, null);
 
         Assert.assertNotNull(mav);
         Assert.assertEquals(UserController.NEW_USER_VIEW, mav.getViewName());
